// src/interfaces/book.ts
export default interface BookInterface {
   id?: number;
   title?: string;
   author?: string;
   date_print?: number;
   copies?: number
   decription?: string
   edition?: number
   is_aviable?: boolean
   last_browed?: number
   owner_id?: number
   pages?: number
   part?: number
   price?: number
   version?: number

}